requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
faker
itsdangerous
email-validator
python-jose
passlib
python-multipart
python-socketio
python-redis
jinja2
aiosmtplib
bcrypt
cloudinary
redis
fastapi-limiter